pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "howdoyoufindme"
version = "0.1.0"
description = "HowDoYouFindMe Backend"
requires-python = ">=3.12"
dependencies = [
    "crewai",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.2",
    "langchain-core>=0.1.4",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.2"
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "asyncio: mark test as async",
]
# Add this line:
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning:litellm.*:",
]

[tool.coverage.run]
source = ["howdoyoufindme"]
omit = [
    "src/howdoyoufindme/cli.py",
    "*/__init__.py",
    "tests/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass"
]
